Running with SecStrAnnotator 2.3, .NET 6.0, Ubuntu 22.04 in Docker

+++ OverProtServer/install.sh
@@ -1,23 +1,25 @@
 #!/bin/bash
+set -e
 
 DIR=$(dirname $0)
 cd $DIR
 
-sudo echo  # check sudo before actually doing something
-
-sudo apt-get update -y
-sudo apt-get install -y python3-venv
-sudo apt-get install -y nginx
-sudo apt-get install -y redis-server
-sudo apt-get install -y gettext-base   # for envsubst
+CLEAN='false'
+SUDO='sudo'
 
 for ARG in $@; do
-    if [ "$ARG" = "--clean" ]; then
-        rm -rf venv/;
-    fi;
+    if [ "$ARG" = "--clean" ]; then CLEAN='true'; fi;
+    if [ "$ARG" = "--no-sudo" ]; then SUDO=''; fi;
 done;
 
+$SUDO apt-get update -y
+$SUDO apt-get install -y --no-install-recommends python3-venv
+$SUDO apt-get install -y --no-install-recommends nginx
+$SUDO apt-get install -y --no-install-recommends redis-server
+$SUDO apt-get install -y --no-install-recommends gettext-base   # for envsubst
+
+if [ "$CLEAN" = "true" ]; then rm -rf venv/; fi;
 python3 -m venv venv/
-source venv/bin/activate
+. venv/bin/activate
 python3 -m pip install -r requirements.txt
 python3 -m pip freeze > requirements-actual.txt
